Add dummy test that copies test file before running actualy test
This ensures that if only the test was changed (i.e. main_test.go is all that was modified), then we need to ensure that it is updated when tests are run so that make test always uses the most up to date tests in the private GOPATH for the build folder.