Adding TEST_PACKAGES arg and adding coverage generation
Now there's an additional argument that can be specified which only works with go 1.10, which is TEST_PACKAGES, and that controls what packages should be generated with coverage information when go test is run. 
This can be used to perform multiple tests and generate multiple package coverage information in the same test, which is useful.
Additionally, now the coverage information is saved inside a file in the GO_COVERAGE directory, and there's a third automatic test which is generated which will generate full HTML output for the coverage statistics from the test run.