Fix bugs with copying non-configured files nested
There was a bug regarding how the files that weren't configured were copied to $GOPATH. The directories were created inside the source and the files were copied there, instead of in the build folder. 
Also set RULE_MESSAGES to off for the copy targets.
Rename this file, as it's not simple anymore.