Run file(generate...) on configured files as well
This allows us to use generator expressions inside files that are to be configured.
Also removing unnecessary strings around some variables.
Lastly, for the package command, we are now more efficient at generate time by keeping track of what directories need to be created, so we only run one cmake command to create any specific directory, whereas before it was one cmake command for every file that was copied.